==> report/1_FB_demographics.ipynb <==
## Descriptive analysis of FB survey demographics

This notebook runs on the full FB survey data (n=1777). Please note that some analyses may run on subsets of this data. <br>
Author: Rosalynn Yang <br>
Date: 11/25/2020

import numpy as np
import pandas as pd

import utility as util

fb = pd.read_csv("../output/fb_numeric.csv")

fb.shape

# manually exclude those <18 yrs old
fb = fb.drop(fb[fb['Q11'] == 1].index)

fb.shape

#### Recode: (1) age

Recode 3, 4, 5, 6 into missing; <br> {male: 1, female: 0} (same as codes used in UAS)

fb['gender'] = np.nan
fb.loc[fb['Q9']==1, ['gender']] = 0
fb.loc[fb['Q9']==2, ['gender']] = 1

fb['gender'].value_counts(dropna=False).sort_index()

#### Recode: (2) marital status

Recode 3 into missing;<br>{married/living w a partner: 1, not married/not living w a partner: 0}

fb['married'] = np.nan
fb.loc[fb['Q10']==1, ['married']] = 1
fb.loc[fb['Q10']==2, ['married']] = 0

fb['married'].value_counts(dropna=False).sort_index()

#### Recode: (3) age group

{19-35 yrs old: 1, 36-65 yrs old: 2, >=66 yrs old: 3}

fb['age_group'] = np.nan
fb.loc[np.isin(fb['Q11'], [2,3]), ['age_group']] = 1
fb.loc[np.isin(fb['Q11'], [4,5,6]), ['age_group']] = 2
fb.loc[fb['Q11']==7, ['age_group']] = 3

fb['age_group'].value_counts(dropna=False).sort_index()

#### Recode: (4) education level

{HS grad or less: 1, Some college: 2, BA or above: 3}

fb['education'] = np.nan
fb.loc[np.isin(fb['Q12'], [1,2]), ['education']] = 1
fb.loc[fb['Q12']==3, ['education']] = 2
fb.loc[np.isin(fb['Q12'], [4,5]), ['education']] = 3

fb['education'].value_counts(dropna=False).sort_index()

#### Crosstabs and chisq tests

Check demographic distributions in the five image conditions <br>
{Control: 1, COVID: 2, Data Privacy: 3, Finance: 4, Mental Health: 5}

fb['Image'].value_counts().sort_index()

demo_cols = ['gender', 'married', 'age_group', 'education']
col_names = ['Control', 'COVID', 'Data Privacy', 'Finance', 'Mental Health']

for col in demo_cols:
    util.crosstab_chisq(col, 'Image', fb, col_names)
