==> report/2_FB_items.ipynb <==
## Descriptive analysis of FB and UAS items

This notebook runs on the full FB survey data (n=1777). Please note that some analyses may run on subsets of this data. <br>
Author: Rosalynn Yang <br>
Date: 11/29/2020

import numpy as np
import pandas as pd
from IPython.display import display, Markdown

import utility as util

fb = pd.read_csv("../output/fb_numeric.csv")
uas = pd.read_csv("../output/uas_numeric.csv")

fb.shape

uas.shape

# manually exclude those <18 yrs old
fb = fb.drop(fb[fb['Q11'] == 1].index)

fb.shape

### Crosstabs of FB items and Image

*Mental health: Q1, Q2 <br>
Data privary: Q3, Q3_1, Q4 <br>
COVID19: Q5, Q6 <br>
Finance: Q8*

fb_items_cols = ['Q1', 'Q2', 'Q3', 'Q3_1', 'Q4', 'Q5', 'Q6', 'Q8']
col_names = ['Neutral(%)', 'COVID(%)', 'Data Privacy(%)', 'Finance(%)', 'Mental Health(%)']

for col in fb_items_cols:
    util.crosstab_chisq(col, 'Image', fb, col_names)

### UAS weighted vs. FB unweighted (all)

fb_uas_cols = ['Q1', 'Q2', 'Q5', 'Q6', 'Q8']
uas_cols = ['cr027a', 'cr027c', 'cr030', 'cr018a', 'ei002']

fb_freq_tables = {}
for col in fb_uas_cols:
    fb_freq_tables[col] = util.freq_percent_table(col, fb)

uas_freq_tables = {}
for col in uas_cols:
    uas_freq_tables[col] = util.weighted_freq_percent_table(col, 'final_weight', uas)

comparison_tables = {}
for fb_uas_col, uas_col in zip(fb_uas_cols, uas_cols):
    temp_df = pd.concat([uas_freq_tables[uas_col], fb_freq_tables[fb_uas_col]], axis=1, ignore_index=True)
    temp_df.columns = ['UAS Sample (%)', 'Facebook Ad Image Sample (%)']
    comparison_tables[fb_uas_col] = temp_df

for fb_uas_col in fb_uas_cols:
    display(Markdown(f"#### Comparing responses from {fb_uas_col}"))
    display(comparison_tables[fb_uas_col])

### FB pooled vs. FB(control)

fb['Image'].value_counts().sort_index()

fb_control = fb.loc[fb['Image']==1,:]
fb_others= fb.loc[fb['Image']>1,:]

fb_control_freq_tables = {}
for col in fb_uas_cols:
    fb_control_freq_tables[col] = util.freq_percent_table(col, fb_control)

fb_others_freq_tables = {}
for col in fb_uas_cols:
    fb_others_freq_tables[col] = util.freq_percent_table(col, fb_others)

fb_comparison_tables = {}
for fb_uas_col in fb_uas_cols:
    temp_df = pd.concat([fb_others_freq_tables[fb_uas_col], fb_control_freq_tables[fb_uas_col]], axis=1, ignore_index=True)
    temp_df.columns = ['All Other Images (%)', 'Neutral Image (%)']
    fb_comparison_tables[fb_uas_col] = temp_df

for fb_uas_col in fb_uas_cols:
    display(Markdown(f"#### Comparing responses from {fb_uas_col}"))
    display(fb_comparison_tables[fb_uas_col])

### UAS weighted demographics

uas_demographics = ['gender', 'maritalstatus', 'age_cat', 'education']

uas_demo_tables = {}
for col in uas_demographics:
    uas_demo_tables[col] = util.weighted_freq_percent_table(col, 'final_weight', uas)

for col in uas_demographics:
    display(Markdown(f"#### Weighted {col} from UAS"))
    display(uas_demo_tables[col])
