==> report/3_FB_raking_prep.ipynb <==
## Prep for raking: FB data

Goal is to prep the data to be ready for raking with the four parameters: 1) age group, 2) gender, 3) marital status, 4) education level <br>
Need to remove missing data from the raking parameters and output 5 separate dataframes for raking

import numpy as np
import pandas as pd

fb = pd.read_csv('../output/fb_numeric.csv')

fb.shape

fb.columns

fb = fb[['Finished', 'RecordedDate', 'Q1', 'Q2', 'Q3', 'Q3_1', 'Q4', 'Q5', 'Q6',
         'Q7_1', 'Q8', 'Q9', 'Q9_6_TEXT', 'Q10', 'Q11', 'Q12', 'Week', 'Image']]

fb.columns = ['Finished', 'RecordedDate', 'Q1', 'Q2', 'Q3', 'Q3_1', 'Q4', 'Q5', 'Q6',
              'Q7_1', 'Q8', 'gender', 'gender_text', 'marital', 'age_group', 'education', 'Week', 'Image']

fb['Finished'].value_counts()

# remove breakoffs
fb_complete = fb.loc[fb['Finished']==True, :]

demographic_cols = ['gender', 'marital', 'age_group', 'education']
for col in demographic_cols:
    print(fb_complete[col].value_counts(dropna=False).sort_values())

*Data cleaning strategy: <br>
'gender': keep only values 1 and 2 and remove missing <br>
'marital': keep only values 1 and 2 and remove missing <br>
'age_group': remove value 1 and missing <br>
'education': figure out what correlates with high missing, keep for now, may be not usable for raking*

# check missing data in education
fb_educ_missing = fb_complete[['education', 'RecordedDate']].fillna(99)

# education data consistently missing for the first week: 2020-05-14 to 2020-05-20
pd.crosstab(fb_educ_missing['education'], fb_educ_missing['RecordedDate'])

# removed rows where gender, marital, and age_group are missing
fb_dropna = fb_complete.dropna(subset=['gender', 'marital', 'age_group'])
fb_dropna.shape

# clean up aforementioned values
fb_cleandemo = fb_dropna[(fb_dropna['gender'] <3) & (fb_dropna['marital']<3) & (fb_dropna['age_group']>=2)]

# output 5 datasets for weighting
dflist = list(fb_cleandemo.groupby('Image'))

condition_codebook = {1: 'control',
                      2: 'covid',
                      3: 'privacy',
                      4: 'finance',
                      5: 'mental'}

condition_dfs = {}

for image_number, condition_df in dflist:
    image_text = condition_codebook[image_number]
    condition_dfs[image_text] = condition_df

# check data shape
for image_text, condition_df in condition_dfs.items():
    print(image_text)
    print(condition_df.shape)

#for image_text, condition_df in condition_dfs.items():
    #for col in demographic_cols:
        #print(image_text)
        #print(condition_df[col].value_counts(dropna=False).sort_values())

# output 5 datasets
for image_text, condition_df in condition_dfs.items():
    condition_df.to_csv(f'../output/{image_text}.csv', index=False)
