==> report/4_UAS_raking_prep.ipynb <==
## Prep for raking: UAS data

Goal is to prep the UAS demos to be consistent with FB levels <br>
Target raking variables: gender, age_group, marital_status

import numpy as np
import pandas as pd

uas = pd.read_csv("../output/uas_numeric.csv")

# recode gender to be consistent with fb
uas['gender_r'] = 1
uas.loc[uas['gender']==1, 'gender_r'] = 2

# recode maritalstatus to be consistent with fb
uas['maritalstatus_r'] = np.nan
uas.loc[np.isin(uas['maritalstatus'], [1, 2]), ['maritalstatus_r']] = 1
uas.loc[np.isin(uas['maritalstatus'], [3, 4, 5, 6]), ['maritalstatus_r']] = 2

# recode education to be consistent with fb
uas['education_r'] = np.nan
uas.loc[np.isin(uas['education'], [1, 2, 3, 4, 5, 6, 7, 8]), ['education_r']] = 1
uas.loc[uas['education']==9, ['education_r']] = 2
uas.loc[np.isin(uas['education'], [10, 11, 12]), ['education_r']] = 3
uas.loc[uas['education']==13, ['education_r']] = 4
uas.loc[np.isin(uas['education'], [14, 15, 16]), ['education_r']] = 5

uas = uas[["uasid", "cr027a", "cr027c", "cr030", "cr018a", "lr015", "ei002", "cr015c", "cr015d", 
           "cr015k", "cr015l", "gender_r", "maritalstatus_r", "age_group", "education_r", "final_weight"]]
uas.columns = ["uasid", "cr027a", "cr027c", "cr030", "cr018a", "lr015", "ei002", "cr015c", "cr015d", 
               "cr015k", "cr015l", "gender", "marital", "age_group", "education", "final_weight"]

# removed rows where gender, marital, and age_group are missing
uas = uas.dropna(subset=['gender', 'marital', 'age_group'])
uas.shape

uas.to_csv('../output/uas_democlean.csv', index=False)
