==> report/6_weighted_freq.ipynb <==
## Weighted frequency distributions and comparisons

This notebook runs on data previously weighted in R using the survey package (survey:::rake)

##### Raking parameters: gender, age group, marital status, education level 
Population marginal distributions based on 2018 ACS 5-Year Estimates - Public Use Microdata <br>

* Gender:
  + 1 = female (51.4%)
  + 2 = male (48.6%) <br><br>
  
* Age group:
  + 2 = 19-25 years old (12.7%) <br>
  + 3 = 26-35 years old (18.1%) <br>
  + 4 = 36-45 years old (16.6%) <br>
  + 5 = 46-55 years old (17.5%) <br>
  + 6 = 56-65 years old (16.5%) <br>
  + 7 = 66+ years old (18.6%) <br><br>
  
* Marital status:
  + 1 = married (48.7%)
  + 2 = never married, divorced, widowed, separated (51.3%) <br><br>

* Education level:
  + 1 = less than high school (12.0%)
  + 2 = high school diploma (27.5%)
  + 3 = some college (31.1%)
  + 4 = bachelor's degree (18.6%)
  + 5 = graduate degree (10.8%) <br>

*Note: <br>
1) can't use education level due to high level of missing data from week1 <br>
2) age <19 is excluded from the population estimates before raking to match the survey data*

import numpy as np
import pandas as pd
from IPython.display import display, Markdown

import utility as util

import warnings
warnings.filterwarnings("ignore")

control = pd.read_csv('../output/control_weighted.csv')
covid = pd.read_csv('../output/covid_weighted.csv')
privacy = pd.read_csv('../output/privacy_weighted.csv')
finance = pd.read_csv('../output/finance_weighted.csv')
mental = pd.read_csv('../output/mental_weighted.csv')
uas = pd.read_csv('../output/uas_weighted.csv')

# Q1, Q2=mental health; Q5, Q6=COVID19; Q8=finance
fb_cols = ['Q1', 'Q2', 'Q5', 'Q6', 'Q8']
uas_cols = ['cr027a', 'cr027c', 'cr030', 'cr018a', 'ei002']

### Condition: control

util.create_comparison_tables(control, uas, fb_cols, uas_cols)

### Condition: COVID

Q5: How likely are you to get vaccinated for coronavirus once a vaccine is available to the public? 1 = Very Unlikely; 2 = Somewhat Unlikely; 3 = Somewhat Likely; 4 = Very Likely; 5 = Unsure <br>
Q6: Are Federal, state, or local governments currently requiring you to limit non-essential travel? 1 = Yes; 2 = No; 3 = Unsure

util.create_comparison_tables(covid, uas, fb_cols, uas_cols)

### Condition: privacy

util.create_comparison_tables(privacy, uas, fb_cols, uas_cols)

### Condition: finance

Q8: In the past seven days, were you worried you would run out of food because of a lack of money or other resources? 1 = Yes; 2 = No; 3 = Unsure

util.create_comparison_tables(finance, uas, fb_cols, uas_cols)

### Condition: mental health

Q1: In the past fourteen days, how often have you felt nervous, anxious, or on edge? 1 = Not at all or less than 1 day; 2 = 1-2 days; 3 = 3-4 days; 4 = 5-7 days <br>
Q2: In the past 7 days, how often have you felt depressed? 1 = Not at all or less than 1 day; 2 = 1-2 days; 3 = 3-4 days; 4 = 5-7 days

util.create_comparison_tables(mental, uas, fb_cols, uas_cols)
