==> report/8.1_SD_daoust_rep_data.ipynb <==
## Social desirability experiment: replicating Daoust findings - data cleaning

Purpose of this notebook: clean up raw data and output analysis datasets

Experiment specifications: <br>
* Daoust experiment:
  + Have you done? Yes=1/No=2/Unsure=3 (Q1.3, Q1.4, Q1.5) <br>
  + Some people XX have you done? Yes=1/Only when neccessary=2/No=3/Unsure=4 (Q2.3, Q2.4, Q2.5) <br><br>
  
* SD benchmark:
  + Individuals are more to blame? Agree=1/disagree=2 (Q3.1) <br>
  + Social conditions are more to blame? Agree=1/disagree=2 (Q3.2) <br>

import numpy as np
import pandas as pd

import warnings
warnings.filterwarnings("ignore")

rep_exp = pd.read_csv('../input/rep_exp.csv')

rep_exp.shape

rep_exp.columns

# columns: select and rename analysis variables
rep_exp = rep_exp[['Finished', 'Q1.3', 'Q1.4', 'Q1.5', 'Q2.3', 'Q2.4', 'Q2.5', 'Q3.1', 'Q3.2', 
                   'Q5.1', 'Q5.2', 'Q5.3', 'Q5.4']]

rep_exp.columns = ['Finished', 'visit_A', 'over_A', 'outdoors_A', 'visit_B', 'over_B', 'outdoors_B', 'blame_individual', 'blame_social', 
                   'sex', 'marital', 'age_group', 'education']

# rows: remove breakoffs
rep_exp = rep_exp.loc[rep_exp['Finished']==1, :]

rep_exp.shape

# rows: remove unsure -> save as analysis dataset "df"
df = rep_exp[(rep_exp['visit_A'] != 3) & (rep_exp['over_A'] != 3) & (rep_exp['outdoors_A'] != 3) & 
             (rep_exp['visit_B'] != 4) & (rep_exp['over_B'] != 4) & (rep_exp['outdoors_B'] != 4)]

df.shape

# rows: check demographic frequencies - at analysis stage: consider whether 1) remove missing demographics, 2) recode demographics
demographic_cols = ['sex', 'marital', 'age_group', 'education']

for col in demographic_cols:
    print(df[col].value_counts(dropna=False).sort_index())
    print("................")

# rows: remove those that are age<18
df = df.loc[~np.isin(df['age_group'], [1]), :]

df.shape

# reset and add index column "id"
df = df.reset_index(drop=True)
df['id'] = range(1, len(df) + 1)

df.head()

# separate into two analysis datasets: Daoust experiment and SD benchmark experiment, both with demographics
df_dst = df[['id', 'visit_A', 'over_A', 'outdoors_A', 'visit_B', 'over_B', 'outdoors_B', 'sex', 'marital', 'age_group', 'education']]
df_benchmark = df[['id','blame_individual', 'blame_social', 'sex', 'marital', 'age_group', 'education']]

#### Clean up df_dauost experimental conditions

df_dst.head()

grouping = np.array(['A', 'B'])
row_groupings = []

for index, row in df_dst[['visit_A', 'visit_B']].iterrows():
    boolean_list = ~row.isna()
    selected = grouping[boolean_list]
    try:
        value_selected = selected[0]
    except IndexError:
        # when all columns are nan, no grouping is available
        row_groupings.append('NAN')
        continue
    row_groupings.append(value_selected)

df_dst['condition'] = row_groupings

df_dst['condition'].value_counts(dropna=False).sort_index()

df_dst.head()

# recode "No" before mixing up
A_cols = ['visit_A', 'over_A', 'outdoors_A'] # recode 2 into 9
B_cols = ['visit_B', 'over_B', 'outdoors_B'] # recode 3 into 9

for col in A_cols:
    df_dst[col] = df_dst[[col]].replace([2], [9])
    
for col in B_cols:
    df_dst[col] = df_dst[[col]].replace([3], [9])

df_dst['visit'] = df_dst[['visit_A', 'visit_B']].fillna(0).sum(axis=1).astype(int)
df_dst['over'] = df_dst[['over_A', 'over_B']].fillna(0).sum(axis=1).astype(int)
df_dst['outdoors'] = df_dst[['outdoors_A', 'outdoors_B']].fillna(0).sum(axis=1).astype(int)

df_dst = df_dst[['id', 'condition', 'visit', 'over', 'outdoors', 'sex', 'marital', 'age_group', 'education']]

df_dst.head()

df_dst.shape

df_dst.to_csv('../output/df_dst.csv', index=False)

#### Clean up df_benchmark experimental conditions

df_benchmark.head()

pd.crosstab(df_benchmark['blame_individual'], df_benchmark['blame_social'])

grouping = np.array(['A', 'B'])
row_groupings = []

for index, row in df_benchmark[['blame_individual', 'blame_social']].iterrows():
    boolean_list = ~row.isna()
    selected = grouping[boolean_list]
    try:
        value_selected = selected[0]
    except IndexError:
        # when all columns are nan, no grouping is available
        row_groupings.append('NAN')
        continue
    row_groupings.append(value_selected)

df_benchmark['condition'] = row_groupings

df_benchmark['condition'].value_counts(dropna=False).sort_index()

df_benchmark.head()

df_benchmark['blame'] = df_benchmark[['blame_individual', 'blame_social']].fillna(0).sum(axis=1).astype(int)

df_benchmark = df_benchmark[['id', 'condition', 'blame', 'sex', 'marital', 'age_group', 'education']]

df_benchmark.head()

df_benchmark.to_csv('../output/df_benchmark.csv', index=False)
