==> report/8.2_SD_daoust_rep_analysis.ipynb <==
## Social desirability experiment: replicating Daoust findings - data analysis

Experiment specifications: <br>
* Daoust experiment (variables: visit, over, outdoors):
  + Condition A: Have you done? Yes/No/Unsure <br>
  + Condition B: Some people XX have you done? Yes/Only when neccessary/No/Unsure <br><br>
  
* SD benchmark (variables: blame):
  + Condition A: Individuals are more to blame? Agree/disagree <br>
  + Condition B: Social conditions are more to blame? Agree/disagree <br>

import numpy as np
import pandas as pd

import utility as util

#### 1) Daoust experiment

df_dst = pd.read_csv('../output/df_dst.csv')

df_dst.shape

df_dst.head()

# numeric -> character display
SD_cols = ['visit', 'over', 'outdoors']

for col in SD_cols:
    df_dst[col] = df_dst[[col]].replace([1, 2, 9], ["1. Yes", "2. Only when necessary/occasionally", "9. No"])

for col in SD_cols:
    util.crosstab_chisq(col, 'condition', df_dst, chisqtest=False)

# combine top two response options into "yes"
combine_top_two = {"1. Yes": "1. Yes", 
                   "2. Only when necessary/occasionally": "1. Yes", 
                   "9. No": "9. No"}

for col in SD_cols:
    df_dst[f'{col}_r'] = df_dst[col].map(combine_top_two)

recoded_cols = ['visit_r', 'over_r', 'outdoors_r']

for col in recoded_cols:
    print('A: Have you done? Yes/No')
    print('B: Some people XX have you done? Yes/Only when neccessary/No')
    util.crosstab_chisq(col, 'condition', df_dst)

#### 2) Benchmark experiment

df_benchmark = pd.read_csv('../output/df_benchmark.csv')

df_benchmark.shape

df_benchmark.head()

df_benchmark['blame'] = df_benchmark['blame'].replace([1, 2], ["1. Agree", "2. Disagree"])

print('A: Individuals are more to blame? Agree/disagree')
print('B: Social conditions are more to blame? Agree/disagree')

util.crosstab_chisq('blame', 'condition', df_benchmark)
