==> report/8.3_SD_dauost_rep_subgroup.ipynb <==
## Social desirability experiment: replicating Daoust findings - subgroup analysis

Experiment specifications: <br>
* Daoust experiment (variables: visit, over, outdoors):
  + Have you done? Yes/No/Unsure <br>
  + Some people XX have you done? Yes/Only when neccessary/No/Unsure <br><br>
  
* SD benchmark (variables: blame):
  + Individuals are more to blame? Agree/disagree <br>
  + Social conditions are more to blame? Agree/disagree <br>

import numpy as np
import pandas as pd

import utility as util

df_dst = pd.read_csv('../output/df_dst.csv')
df_benchmark = pd.read_csv('../output/df_benchmark.csv')

#### recode demographics

df_dst.columns

demographic_cols = ['sex', 'marital', 'age_group', 'education']

for col in demographic_cols:
    print(df_dst[col].value_counts(dropna=False).sort_index())
    print('>>>>>>>>>>')

sex_map = {1: '1. male',
           2: '2. female'}

marital_map = {1: '1. married',
               2: '2. unmarried',
               3: '2. unmarried',
               4: '2. unmarried',
               5: '2. unmarried'}

age_g_map = {2: '18-34',
             3: '18-34',
             4: '35-54', 
             5: '35-54',
             6: '55+',
             7: '55+'}

education_map ={1: '1. Less than college',
                2: '1. Less than college',
                3: '1. Less than college',
                4: '2. College+',
                5: '2. College+'}

df_dst['sex_r'] = df_dst['sex'].map(sex_map)
df_dst['marital_r'] = df_dst['marital'].map(marital_map)
df_dst['age_group_r'] = df_dst['age_group'].map(age_g_map)
df_dst['education_r'] = df_dst['education'].map(education_map)

df_benchmark['sex_r'] = df_benchmark['sex'].map(sex_map)
df_benchmark['marital_r'] = df_benchmark['marital'].map(marital_map)
df_benchmark['age_group_r'] = df_benchmark['age_group'].map(age_g_map)
df_benchmark['education_r'] = df_benchmark['education'].map(education_map)

recoded_demo_cols = ['sex_r', 'marital_r', 'age_group_r', 'education_r']

for col in recoded_demo_cols:
    print(df_dst[col].value_counts(dropna=False).sort_index())
    print('>>>>>>>>>>')

df_dst.groupby('sex_r').mean()

SD_cols = ['visit', 'over', 'outdoors']

for col in SD_cols:
    df_dst[col] = df_dst[[col]].replace([1, 2, 9], ["1. Yes", "2. Only when necessary/occasionally", "9. No"])

# combine top two response options into "yes"
combine_top_two = {"1. Yes": "1. Yes", 
                   "2. Only when necessary/occasionally": "1. Yes", 
                   "9. No": "9. No"}

for col in SD_cols:
    df_dst[f'{col}_r'] = df_dst[col].map(combine_top_two)

recoded_cols = ['visit_r', 'over_r', 'outdoors_r']

woman_df = df_dst.loc[df_dst['sex'] == 1]
man_df = df_dst.loc[df_dst['sex'] == 2]

woman_df.head()

for col in recoded_cols:
    util.crosstab_chisq(col, 'condition', woman_df)

for col in recoded_cols:
    util.crosstab_chisq(col, 'condition', man_df)
