==> report/9_STL_rep_analysis.ipynb <==
## Analysis for the manuscript: prepare data

This analysis is conducted based on the Dauost experiment replication data, fielded in 2020. <br>
1) checking for randomization of treatment conditions <br>
2) experimental data analysis <br>
3) three-way interactions

import numpy as np
import pandas as pd

import warnings
warnings.filterwarnings("ignore")

rep_exp = pd.read_csv('../input/rep_exp.csv')

# columns: select and rename analysis variables
rep_exp = rep_exp[['Finished', 'Q1.3', 'Q1.4', 'Q1.5', 'Q2.3', 'Q2.4', 'Q2.5', 'Q3.1', 'Q3.2', 
                   'Q5.1', 'Q5.2', 'Q5.3', 'Q5.4']]

rep_exp.columns = ['Finished', 'visit_A', 'over_A', 'outdoors_A', 'visit_B', 'over_B', 'outdoors_B', 'blame_individual', 'blame_social', 
                   'sex', 'marital', 'age_group', 'education']

# rows: remove breakoffs
rep_exp = rep_exp.loc[rep_exp['Finished']==1, :]

rep_exp.shape

# rows: check demographic frequencies - at analysis stage: consider whether 1) remove missing demographics, 2) recode demographics
demographic_cols = ['sex', 'marital', 'age_group', 'education']

for col in demographic_cols:
    print(rep_exp[col].value_counts(dropna=False).sort_index())
    print("................")

# rows: remove those that are age<18
rep_exp = rep_exp.loc[~np.isin(rep_exp['age_group'], [1]), :]

# rows: remove those with missing demographics
rep_exp = rep_exp.dropna( how='any', subset= demographic_cols)

rep_exp.shape

# reset and add index column "id"
rep_exp = rep_exp.reset_index(drop=True)
rep_exp['id'] = range(1, len(rep_exp) + 1)

# extract the replication dataset
df_rep = rep_exp[['id', 'visit_A', 'over_A', 'outdoors_A', 'visit_B', 'over_B', 'outdoors_B', 'sex', 'marital', 'age_group', 'education']]

# recoding for consistency 
A_cols = ['visit_A', 'over_A', 'outdoors_A']
B_cols = ['visit_B', 'over_B', 'outdoors_B'] 

for col in A_cols:
    df_rep[col] = df_rep[[col]].replace([3], [8]) # unsure=8
    
for col in B_cols:
    df_rep[col] = df_rep[[col]].replace([4], [8]) # unsure=8
    
for col in B_cols:
    df_rep[col] = df_rep[[col]].replace([2], [11]) # only when neccessary=11
    
for col in B_cols:
    df_rep[col] = df_rep[[col]].replace([3], [2]) # no=2

# assign conditions
grouping = np.array(['A', 'B'])
row_groupings = []

for index, row in df_rep[['visit_A', 'visit_B']].iterrows():
    boolean_list = ~row.isna()
    selected = grouping[boolean_list]
    try:
        value_selected = selected[0]
    except IndexError:
        # when all columns are nan, no grouping is available
        row_groupings.append('NAN')
        continue
    row_groupings.append(value_selected)

df_rep['condition'] = row_groupings

df_rep['condition'].value_counts(dropna=False).sort_index()

df_rep.head()

# combine two sets of variables into one
df_rep['visit'] = df_rep[['visit_A', 'visit_B']].fillna(0).sum(axis=1).astype(int)
df_rep['over'] = df_rep[['over_A', 'over_B']].fillna(0).sum(axis=1).astype(int)
df_rep['outdoors'] = df_rep[['outdoors_A', 'outdoors_B']].fillna(0).sum(axis=1).astype(int)

df_rep = df_rep[['id', 'condition', 'visit', 'over', 'outdoors', 'sex', 'marital', 'age_group', 'education']]

df_rep.head()

df_rep.shape

# rows: check frequency of experimental variables
experiment_cols = ['visit', 'over', 'outdoors']

for col in experiment_cols:
    print(df_rep[col].value_counts(dropna=False).sort_index())
    print("................")

df_rep.to_csv('../output/df_rep.csv', index=False)
