==> report/9.3_STL_extension_data.ipynb <==
## Analysis based on the extension study data

import numpy as np
import pandas as pd
import utility as util

import warnings
warnings.filterwarnings("ignore")

ext_df = pd.read_csv('../output/SD_experiment_df.csv')

# reset and add index column "id"
ext_df = ext_df.reset_index(drop=True)
ext_df['id'] = range(1, len(ext_df) + 1)

ext_df = ext_df[['id', 'condition', 'gone_to_friend', 'had_visitors', 'had_close_contact', 'gone_outside',
                 'gender', 'marital', 'age_group', 'education']]

# remove missing based on demographic variables
democols = ['gender', 'marital', 'age_group', 'education']

ext_df = ext_df.dropna(how='any', subset= democols)

# remove those that reported unsure in the marital variable
ext_df = ext_df.loc[~np.isin(ext_df['marital'], [8]), :]

ext_df.shape

### 1) checking for randomization of treatment conditions

democols = ['gender', 'marital', 'age_group', 'education']

for col in democols:
    print(ext_df[col].value_counts(dropna=False).sort_index())
    print("................")

# recode demographic variables for consistency
ext_df['gender_r'] = np.nan
ext_df.loc[ext_df['gender']==1, ['gender_r']] = 2 # woman
ext_df.loc[ext_df['gender']==2, ['gender_r']] = 1 # man

ext_df['marital_r'] = np.nan
ext_df.loc[ext_df['marital']==1, ['marital_r']] = 1 # married
ext_df.loc[ext_df['marital']==7, ['marital_r']] = 2 # not married

ext_df['age_group_r'] = np.nan
ext_df.loc[np.isin(ext_df['age_group'], [2, 3, 4, 5]), ['age_group_r']] = 1 # 19-55
ext_df.loc[ext_df['age_group']==6, ['age_group_r']] = 2 # 56-65
ext_df.loc[ext_df['age_group']==7, ['age_group_r']] = 3 # 66+

ext_df['education_r'] = np.nan
ext_df.loc[np.isin(ext_df['education'], [4, 5, 6]), ['education_r']] = 1 # less than college
ext_df.loc[np.isin(ext_df['education'], [7, 8]), ['education_r']] = 2 # college or more

# checking recoded variables
democols_r = ['gender_r', 'marital_r', 'age_group_r', 'education_r']

for col in democols_r:
    print(ext_df[col].value_counts(dropna=False).sort_index())
    print("................")

for col in democols_r:
    util.crosstab_chisq(col, 'condition', ext_df, chisqtest=True)

### 1.5) Control vs. ABC combined

ext_df['condition'].value_counts()

ext_df['condition_r'] = np.nan
ext_df.loc[ext_df['condition']=='Control', ['condition_r']] = 'Control'
ext_df.loc[ext_df['condition']=='A', ['condition_r']] = 'Experiment'
ext_df.loc[ext_df['condition']=='B', ['condition_r']] = 'Experiment'
ext_df.loc[ext_df['condition']=='C', ['condition_r']] = 'Experiment'

ext_df['condition_r'].value_counts()

for col in democols_r:
    util.crosstab_chisq(col, 'condition_r', ext_df, chisqtest=True)

### 2) experimental data analysis

# check missing in experiment cols
expcols = ['gone_to_friend', 'had_visitors', 'had_close_contact', 'gone_outside']

for col in expcols:
    print(ext_df[col].value_counts(dropna=False).sort_index())
    print("................")

# numeric -> character display
for col in expcols:
    ext_df[col] = ext_df[[col]].replace([1, 2, 3, 4], ["1a. Yes/Yes for any reason", "1b. Yes but only when neccessary", 
                                                       "2. No", "3. Unsure"])

# raw distributions for cross-tabs
for col in expcols:
    util.crosstab_chisq(col, 'condition', ext_df, chisqtest=False)

# combining 1a, 1b, and 8 together as 1/yes
combine_three = {"1a. Yes/Yes for any reason": "1. Yes", 
                 "2. No": "2. No",
                 "1b. Yes but only when neccessary": "1. Yes", 
                 "3. Unsure": "1. Yes"}

for col in expcols:
    ext_df[f'{col}_r'] = ext_df[col].map(combine_three)
    
recoded_cols = ['gone_to_friend_r', 'had_visitors_r', 'had_close_contact_r', 'gone_outside_r'] # use recoded_cols for further analyses

ext_df['condition'].value_counts()

# subset data for control vs. particular treatment condition comparisons
ext_A = ext_df[ext_df['condition'].isin(['Control', 'A'])]
ext_B = ext_df[ext_df['condition'].isin(['Control', 'B'])]
ext_C = ext_df[ext_df['condition'].isin(['Control', 'C'])]

# cross-tabs based on recodes
for col in recoded_cols:
    util.crosstab_chisq(col, 'condition', ext_A, chisqtest=True)

# cross-tabs based on recodes
for col in recoded_cols:
    util.crosstab_chisq(col, 'condition', ext_B, chisqtest=True)

# cross-tabs based on recodes
for col in recoded_cols:
    util.crosstab_chisq(col, 'condition', ext_C, chisqtest=True)

### 3) testing interactions

ext_df['condition_r'] = np.nan
ext_df.loc[ext_df['condition']=='Control', ['condition_r']] = 1 # Control
ext_df.loc[ext_df['condition']=='A', ['condition_r']] = 2 # condition A
ext_df.loc[ext_df['condition']=='B', ['condition_r']] = 3 # condition B
ext_df.loc[ext_df['condition']=='C', ['condition_r']] = 4 # condition B

# recode y as numerics
for col in recoded_cols:
    ext_df[col] = ext_df[[col]].replace(['1. Yes'], [1]) # yes have done the mentioned behavior
    ext_df[col] = ext_df[[col]].replace(['2. No'], [0]) # no haven not done the mentioned behavior

ext_df['condition_r'].value_counts()

recoded_cols

# use R for logistic regression tables
%load_ext rpy2.ipython

%%R -i ext_df
print('GONE TO FRIEND - SEX')
summary(glm(gone_to_friend_r ~ factor(condition_r) * factor(gender_r), family=binomial, data=ext_df))

%%R -i ext_df
print('GONE TO FRIEND - MARITAL')
summary(glm(gone_to_friend_r ~ factor(condition_r) * factor(marital_r), family=binomial, data=ext_df))

%%R -i ext_df
print('GONE TO FRIEND - AGE_GROUP')
summary(glm(gone_to_friend_r ~ factor(condition_r) * factor(age_group_r), family=binomial, data=ext_df))

%%R -i ext_df
print('GONE TO FRIEND - EDUCATION')
summary(glm(gone_to_friend_r ~ factor(condition_r) * factor(education_r), family=binomial, data=ext_df))

%%R -i ext_df
print('HAD VISITORS - SEX')
summary(glm(had_visitors_r ~ factor(condition_r) * factor(gender_r), family=binomial, data=ext_df))

%%R -i ext_df
print('HAD VISITORS - MARITAL')
summary(glm(had_visitors_r ~ factor(condition_r) * factor(marital_r), family=binomial, data=ext_df))

%%R -i ext_df
print('HAD VISITORS - AGE_GROUP')
summary(glm(had_visitors_r ~ factor(condition_r) * factor(age_group_r), family=binomial, data=ext_df))

%%R -i ext_df
print('HAD VISITORS - EDUCATION')
summary(glm(had_visitors_r ~ factor(condition_r) * factor(education_r), family=binomial, data=ext_df))

%%R -i ext_df
print('HAD CLOSE CONTACT - SEX')
summary(glm(had_close_contact_r ~ factor(condition_r) * factor(gender_r), family=binomial, data=ext_df))

%%R -i ext_df
print('HAD CLOSE CONTACT - MARITAL')
summary(glm(had_close_contact_r ~ factor(condition_r) * factor(marital_r), family=binomial, data=ext_df))

%%R -i ext_df
print('HAD CLOSE CONTACT - AGE_GROUP')
summary(glm(had_close_contact_r ~ factor(condition_r) * factor(age_group_r), family=binomial, data=ext_df))

%%R -i ext_df
print('HAD CLOSE CONTACT - EDUCATION')
summary(glm(had_close_contact_r ~ factor(condition_r) * factor(education_r), family=binomial, data=ext_df))

%%R -i ext_df
print('GONE_OUTSIDE - SEX')
summary(glm(gone_outside_r ~ factor(condition_r) * factor(gender_r), family=binomial, data=ext_df))

%%R -i ext_df
print('GONE_OUTSIDE - MARITAL')
summary(glm(gone_outside_r ~ factor(condition_r) * factor(marital_r), family=binomial, data=ext_df))

%%R -i ext_df
print('GONE_OUTSIDE - AGE_GROUP')
summary(glm(gone_outside_r ~ factor(condition_r) * factor(age_group_r), family=binomial, data=ext_df))

%%R -i ext_df
print('GONE_OUTSIDE - EDUCATION')
summary(glm(gone_outside_r ~ factor(condition_r) * factor(education_r), family=binomial, data=ext_df))

# checking interaction for the one significant main effect: control vs. C, had_visitors
ext_C = ext_df[ext_df['condition'].isin(['Control', 'C'])]

%%R -i ext_C
print('HAD_VISITORS - SEX')
summary(glm(had_visitors_r ~ factor(condition_r) * factor(gender_r), family=binomial, data=ext_C))

%%R -i ext_C
print('HAD_VISITORS - MARITAL')
summary(glm(had_visitors_r ~ factor(condition_r) * factor(marital_r), family=binomial, data=ext_C))

%%R -i ext_C
print('HAD_VISITORS - AGE_GROUP')
summary(glm(had_visitors_r ~ factor(condition_r) * factor(age_group_r), family=binomial, data=ext_C))

%%R -i ext_C
print('HAD_VISITORS - EDUCATION')
summary(glm(had_visitors_r ~ factor(condition_r) * factor(education_r), family=binomial, data=ext_C))
